Replace user info with a popover
